requirements.txt
dj-database-url==0.4.2
Django==1.11.5
djangorestframework==3.6.4
pbr==3.1.1
pytz==2017.2
six==1.10.0
stevedore==1.24.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
